Organize UI components in project structure
