+++ src/Root.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
 import Layout from './screens/Layout';
-import ScrollToTop from './components/ScrollToTop';
+import ScrollToTop from './components/UI/ScrollToTop';
 // eslint-disable-next-line
 import FontAwesomeContainer from './screens/FontAwesomeContainer';
