+++ src/components/UI/FeatureBox.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import { chopString } from '../services/helpers';
+import { chopString } from '../../services/helpers';
 
 const FeatureBox = props => {
   const { description, id, image, label } = props.item;
