+++ src/components/Work/Work.js
@@ -2,11 +2,11 @@ import React, { Component } from 'react';
 import { withRouter } from 'react-router';
 import { connect } from 'react-redux';
 import * as elasticsearchApi from '../../api/elasticsearch-api.js';
-import ErrorSection from '../../components/ErrorSection';
+import ErrorSection from '../UI/ErrorSection';
 import ItemDetail from '../../components/Work/ItemDetail';
 import * as elasticsearchParser from '../../services/elasticsearch-parser';
 import * as globalVars from '../../services/global-vars';
-import LoadingSpinner from '../../components/LoadingSpinner';
+import LoadingSpinner from '../UI/LoadingSpinner';
 import { shuffleArray } from '../../services/helpers';
 import ParentCollections from '../../components/Work/ParentCollections';
 import LargeFeature from '../../components/Work/LargeFeature';
